+++ GHEtool/Examples/main_functionalities.py
@@ -92,7 +92,7 @@ def main_functionalities():
     # disable the use of constant_Rb with the setup, in order to plot the profile correctly
     # when it is given as an argument to the size function, it will size correctly, but the plot will be with
     # constant Rb* since it has not been changed in the setup function
-    borefield.sizing_setup(use_constant_Rb=False)
+    borefield.calculation_setup(use_constant_Rb=False)
     depth = borefield.size()
     print("The borehole depth is: ", str(round(depth, 2)), "m for a sizing with dynamic Rb*.")
     borefield.print_temperature_profile(legend=True)