+++ GHEtool/gui/data_2_borefield_func.py
@@ -95,13 +95,13 @@ def data_2_borefield(ds: DataStorage) -> tuple[Borefield, partial[[], None]]:
             borefield.load = hourly_data
 
     # set up the borefield sizing
-    borefield.sizing_setup(use_constant_Rb=ds.option_method_rb_calc == 0,
-                           L2_sizing=ds.option_method_size_depth == 0,
-                           L3_sizing=ds.option_method_size_depth == 1,
-                           L4_sizing=ds.option_method_size_depth == 2,
-                           atol=ds.option_atol,
-                           rtol=ds.option_rtol,
-                           max_nb_of_iterations=ds.option_max_nb_of_iter)
+    borefield.calculation_setup(use_constant_Rb=ds.option_method_rb_calc == 0,
+                                L2_sizing=ds.option_method_size_depth == 0,
+                                L3_sizing=ds.option_method_size_depth == 1,
+                                L4_sizing=ds.option_method_size_depth == 2,
+                                atol=ds.option_atol,
+                                rtol=ds.option_rtol,
+                                max_nb_of_iterations=ds.option_max_nb_of_iter)
 
     # set borefield simulation period
     borefield.simulation_period = ds.option_simu_period